% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\docType{data}
\name{HELP.FILE.HANDLE}
\alias{HELP.FILE.HANDLE}
\title{lx binary file handle}
\description{
a file \code{handle} (called handle for short) 
is an R object (see notes) wrapping a binary file connection.
(do not confuse with a \link{HELP.FILE.HANDLER})\cr
It is intended to provide a uniform interface for different
file types within \bold{lx} and its extensions.

a \code{handle} contains four mandatory fields:
\itemize{
 \item filename : a character string containing the file name (or url)
 \item type     : a character string containing the file content type (default is "none")
 \item mode     : a character string containing the file opening mode in \link{lx.open}
 \item connect  : an R object containing the file physical connection, the type depends upon file type
 \item handler  : a file handler object. see \link{HELP.FILE.HANDLER}
}

and some optional fields:
\itemize{
 \item header   : file specific information (header, e.g. index table)
}

\code{handle}s are openened and closed thru
\link{lx.open} and \link{lx.close} functions.\cr

files are opened as binary and lx provides the following
binary IO functions to read fixed length bytes from them.
\itemize{
 \item \link{lx.read.int8}  : 8 bits integer(s)
 \item \link{lx.read.int16} : 16 bits integer(s)
 \item \link{lx.read.int32} : 32 bits integer(s)
 \item \link{lx.read.int64} : 64 (actually 53) bits integer(s)
 \item \link{lx.read.string} : basta string
}
}
\note{
in practice a FILE.HANDLE is currently implemented as an R environment

no write operations yet... time is short.
}
