% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.globals.r
\docType{data}
\name{HELP.LX.OPTIONS}
\alias{HELP.LX.OPTIONS}
\title{lx options}
\description{
\subsection{Lx options}{
the following options can be accessed and modified by \link{lx.options}
\tabular{lll}{
  \bold{name} \tab \bold{descr.} \tab \bold{default}\cr
  verbose \tab be verbose \tab TRUE\cr
  use.threads \tab use multithreading \tab FALSE\cr
  mc.cores \tab number of cores to use \tab parallel::detectCores()\cr
  pg.verbose \tab \link{lx.lapply} use progress bar \tab TRUE\cr
  pg.options.style \tab progress bar style \tab 1\cr
  pg.options.time \tab add system.time in progress bar \tab TRUE\cr
  tex.dir \tab temporary dir for tex files \tab tmp\cr
  tex.docclass \tab class of tex document \tab report\cr
  tex.packages \tab latex packages \tab tmp\cr
  tex.graphics.driver \tab tex graphics driver \tab pdf\cr
  tex.driver.options \tab graphics driver options \tab NULL\cr
  tex.graphics.ext \tab tex graphics file extension \tab NULL\cr
}}
note: \code{tex.driver.options} depends upon \code{tex.graphics.driver} you
should usually change them together. e.g. with \code{tex.graphics.driver=="jpeg"}
you may set \code{tex.driver.options} to \code{list(units="in", res=600)}
}
