% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.apply.r
\name{lx.apply}
\alias{lx.apply}
\title{apply wrapper}
\usage{
lx.apply(X, MARGIN, FUN, ..., pg.verbose = lx.options("pg.verbose"))
}
\arguments{
\item{X}{an array, including a matrix. see \link{apply}.}

\item{MARGIN}{a vector giving the subscripts which the
function will be applied over. see \link{apply}.}

\item{FUN}{the function to be applied. see \link{apply}.}

\item{...}{anything passed to FUN. see \link{apply}.}

\item{pg.verbose}{use progress bar}
}
\description{
\code{lx.apply} will call different \link{apply} flavors
depending upon the \code{pg.verbose}.argument:\cr
if pg.verbose is TRUE then\cr
__use \link{apply} with progress bar\cr
else\cr
__use \link{apply} without progress bar\cr
}
\note{
default value for  \code{pg.verbose} is taken from \link{lx.options}
and may therefore be controlled globally.

there is no multithread flavor yet, see \link{lx.rowapply}, \link{lx.lapply}
}
\examples{
n <- 100
lx.options(pg.verbose=TRUE)
system.time(x <- lx.apply(matrix(1:n, ncol=2), 1, function(x) sum(rnorm(5000*sum(x)))))
lx.options(pg.verbose=FALSE)
system.time(x <- lx.apply(matrix(1:n, ncol=2), 1, function(x) sum(rnorm(5000*sum(x)))))

}
