% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.barplot}
\alias{lx.barplot}
\title{barplot with values indicated on top of bars}
\usage{
lx.barplot(x, col = NULL, col.text = NULL, cex.text = 1,
  pos.text = NULL, off.text = 0.5, srt.text = 0, ...)
}
\arguments{
\item{x}{vector or matrix of values}

\item{col}{bar colors (see \link{barplot})}

\item{col.text}{text color (or NULL to select contrast color)}

\item{cex.text}{text magnification}

\item{pos.text}{text position (see \link{text}, NULL means autopos)}

\item{off.text}{text offset (see \link{text})}

\item{srt.text}{text rotation (in degrees)}

\item{...}{arguments to \link{barplot}}
}
\value{
same as \link{barplot}
}
\description{
barplot with values indicated on top of bars
}
\details{
if \code{pos.text} is NULL then the label position will be
choosen to fit within the plot.\cr
if \code{col.text} is NULL then the label color will be
choosen to fit with the bar color.\cr
}
\note{
\code{pos.text} NULL does not work well if srt.text != 0
}
\examples{
\dontrun{
lx.barplot(c(apple=10, orange=5, banana=15))
lx.barplot(c(apple=10, orange=5, banana=15), col=lx.COLORS)
lx.barplot(c(apple=10, orange=5, banana=15), col.text=0, pos.text=1, col=lx.COLORS)
}
}
