% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.close}
\alias{lx.close}
\title{close file handle}
\usage{
lx.close(handle)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}
}
\value{
handle (modified)
}
\description{
close file handle
}
\note{
you should reassign handle on return since
some internal fields have been modified
}
