% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.color.change}
\alias{lx.color.change}
\title{change color(s) hue, saturation or value}
\usage{
lx.color.change(color, dh = 0, ds = 0, dv = 0)
}
\arguments{
\item{color}{vector of any of the three kinds of R color specifications, 
i.e., either a color name (as listed by \link{colors}()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see \link{rgb}),
or a positive integer i meaning \link{palette}()[i]}

\item{dh}{change in hue (-1 <= dh <= 1)}

\item{ds}{change in saturation (-1 <= ds <= 1)}

\item{dv}{change in value (-1 <= dv <= 1)}
}
\value{
modified colors (as hexadecimal strings)
}
\description{
change color(s) hue, saturation or value
}
\examples{
lx.color.change("#ff0000", ds=-0.5)  # => #ff8080
lx.color.change("red", ds=-0.5)      # => #ff8080
}
\seealso{
\link{lx.color.light}, \link{lx.color.alpha}
}
