% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.color.deltaE}
\alias{lx.color.deltaE}
\title{DeltaE distance between two colors}
\usage{
lx.color.deltaE(col1, col2)
}
\arguments{
\item{col1, col2}{colors any of the three kinds of R color specifications, 
i.e., either a color name (as listed by \link{colors}()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see \link{rgb}),
or a positive integer i meaning \link{palette}()[i]}
}
\value{
positive numeric DeltaE
}
\description{
compute the DeltaE distance between two colors
according to the CIE 1976 definition.
}
\examples{
lx.color.deltaE("white", "white")
lx.color.deltaE("white", "red")
}
