% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.color.light}
\alias{lx.color.light}
\title{lighten or darken color(s)}
\usage{
lx.color.light(color, delta)
}
\arguments{
\item{color}{vector of any of the three kinds of R color specifications, 
i.e., either a color name (as listed by \link{colors}()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see \link{rgb}),
or a positive integer i meaning \link{palette}()[i]}

\item{delta}{amount in percentage (-100 <= delta <= 100)\cr
delta > 0 : lighten; delta < 0 : darken}
}
\value{
modified colors (as hexadecimal strings)
}
\description{
lighten or darken color(s)
}
\examples{
lx.color.light("red", 50)  # => #ff8080
lx.color.light("red", -50) # => #800000
}
\seealso{
\link{lx.color.change}, \link{lx.color.alpha}, \link{lx.color.lum}
}
