% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.globals.r
\docType{data}
\name{lx.COLORS}
\alias{lx.COLORS}
\alias{lx.RED}
\alias{lx.GREEN}
\alias{lx.BLUE}
\alias{lx.GREY}
\alias{lx.LGREY}
\alias{lx.TRANSP}
\title{lx global colors}
\description{
\subsection{Lx global colors}{
\tabular{ll}{
  \bold{name} \tab \bold{descr.}\cr
  lx.COLORS \tab a nice looking color scale from d3.scale.category10\cr
  lx.RED  \tab semi-transparent red color for plots\cr
  lx.GREEN  \tab semi-transparent green color for plots\cr
  lx.BLUE  \tab semi-transparent blue color for plots\cr
  lx.GREY  \tab semi-transparent grey color for plots\cr
  lx.LGREY  \tab semi-transparent lightgrey color for plots\cr
  lx.TRANSP  \tab semi-transparent white color for plots\cr
}}
}
\seealso{
\link{lx.rainbow}
}
