% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.algebra.r
\name{lx.crossprod}
\alias{lx.crossprod}
\title{generalized matrix cross-product}
\usage{
lx.crossprod(x, y = NULL, FUN, ...)
}
\arguments{
\item{x}{nxm numerical matrix}

\item{y}{mxn numerical matrix (if NULL then y = t(x))}

\item{FUN}{function called for each row and col as \code{FUN(row, col,...)}}

\item{...}{other parameters to FUN}
}
\value{
nxn numerical matrix. where entry (i,j) is \code{FUN(x[i,], y[,j],...)}
}
\description{
compute cross-product of conformable matrices using any function
of row and col.
}
\note{
usual matrix crosproduct corresponds to \code{FUN=function(x, y) sum(x*y)}
(see examples)
}
\examples{
x <- matrix(1:10, ncol=2)
# usual crossproduct (not efficient)
lx.crossprod(t(x), x, function(x, y) sum(x*y))
# this is the same as (more efficient):
t(x) \%*\% x
# and the same as (more efficient):
crossprod(x)
# cross product with max
lx.crossprod(t(x), FUN=function(x, y) max(x*y))
}
\seealso{
\link{crossprod}
}
