% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.decimate}
\alias{lx.decimate}
\title{decimate vector}
\usage{
lx.decimate(x, length.out, as.index = FALSE)
}
\arguments{
\item{x}{vector}

\item{length.out}{desired length of the resulting vector.}

\item{as.index}{if TRUE return indices else returns values}
}
\description{
downsample vector to length \code{length.out} by selecting
equally spaced indices from 1 to \code{length(x)}.
}
\note{
if \code{length(x) <= length.out} then the original
vector x (or indices) is returned.
}
\examples{
lx.decimate(1:100, 10)
lx.decimate(letters, 5)
lx.decimate(letters, 5, as.index=TRUE)
}
