% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.default.file.handler}
\alias{lx.default.file.handler}
\title{create a default file handler}
\usage{
lx.default.file.handler(name)
}
\arguments{
\item{name}{handler name}
}
\description{
the default file handler handles usual R
binary files.
}
\details{
a default handler (with name "default") is always 
added at the lowest position in handlers stack. So, usually
you don't need to call this function. It is mostly intented
for the case where you want to design your own handler, not
starting from scratch. A typical case is to create a default
R file handler and to override the \code{accept} function to
specify which files to accept.
}
