% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.density}
\alias{lx.density}
\title{Kernel Density Estimation}
\usage{
lx.density(x, bw = "nrd0", ...)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{bw}{the smoothing bandwidth to be used (see \link{density})}

\item{...}{any other parameter to \link{density}}
}
\value{
an object of class "density" (see \link{density})
}
\description{
this is a wrapper around \link{density} that workarounds a (known) bug
with bw="SJ".  The density function generates an error when bw="SJ"
and length(x) > 46341. The message is:\cr
\code{Error in bw.SJ(x, method = "ste") : sample is too sparse to find TD}\cr
the workaround consists in sampling 46341 values.
}
\note{
the seed for sampling is always the same and therefore
the procedure remains deterministic.
}
\examples{
\dontrun{
x <- lx.density(rnorm(50000), "SJ")
}

}
