% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.dup.handle}
\alias{lx.dup.handle}
\title{duplicate file handle}
\usage{
lx.dup.handle(handle)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}
}
\value{
duplicated file handle
}
\description{
duplicate file handle with a new physical connection
}
\details{
this is mostly used in multithreading when accessing the same file
(in read mode) by multiple threads.
}
