% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.erase}
\alias{lx.erase}
\title{erase previously saved or serialized object}
\usage{
lx.erase(name)
}
\arguments{
\item{name}{of first object used in \code{lx.save} or \code{lx.serialize}}
}
\description{
erase previously saved or serialized object
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.save(tab)
lx.erase("tab")
}
\seealso{
\code{\link{lx.save} \link{lx.serialize}}
}
