% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.false}
\alias{lx.false}
\title{always FALSE function}
\usage{
lx.false(...)
}
\arguments{
\item{...}{anything (ignored arguments)}
}
\value{
FALSE
}
\description{
this function always returns FALSE, whatever
the arguments.
}
