% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.osutil.r
\name{lx.file.ext}
\alias{lx.file.ext}
\title{filename extension}
\usage{
lx.file.ext(path)
}
\arguments{
\item{path}{filename}
}
\value{
file (name) extension (excluding the leading dot)
}
\description{
filename extension
}
\note{
only purely alphanumeric extensions are recognized

borrowed from package tools
}
\examples{
lx.file.ext("dir/file.more.ext")
}
