% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.osutil.r
\name{lx.file.no.ext}
\alias{lx.file.no.ext}
\title{filename without extension}
\usage{
lx.file.no.ext(path)
}
\arguments{
\item{path}{filename}
}
\value{
filename without extension (and the leading dot)
}
\description{
filename without extension
}
\note{
only purely alphanumeric extensions are recognized

borrowed from package tools
}
\examples{
lx.file.no.ext("dir/file.more.ext")
}
