% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.osutil.r
\name{lx.file.resolve}
\alias{lx.file.resolve}
\title{resolve filename}
\usage{
lx.file.resolve(path, ext = "", mode = 0, .verbose = FALSE,
  .dedots = TRUE)
}
\arguments{
\item{path}{filename (see details)}

\item{ext}{a character vector or list of character vectors containing
extensions to be combined (see details).}

\item{mode}{access mode (see \link{file.access})}

\item{.verbose}{printout filenames trials}

\item{.dedots}{remove possible double dots in filename
(this may occur if you have included "" as an extension (see examples)).}
}
\value{
actual pathname if a file has matched else NULL
}
\description{
resolve filename by trying different
(combinations of) extensions.
}
\details{
if \code{path} has the form "sys:package:" then
the file is searched in package system files.\cr\cr
if \code{ext} is a list of character vectors then
extensions from each vector element are catenated 
in a 'depth first' way i.e. first element of the first vector
with first of the second ... etc (see examples).\cr
}
\note{
only the first matching filename (if any) is returned.
}
\examples{
lx.file.resolve("sys:lx:data/lx.iris", c("rds", "RData"))
lx.file.resolve("none", .verbose=TRUE)
lx.file.resolve("none", c("", "a", "b"), .verbose=TRUE)
lx.file.resolve("none", list(c("a", "b"), c("", "c", "d"), c("", "e", "f")), .verbose=TRUE)
}
