% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.apply.r
\name{lx.Filter}
\alias{lx.Filter}
\title{Filter wrapper}
\usage{
lx.Filter(FUN, X, pg.verbose = lx.options("pg.verbose"),
  use.threads = lx.use.threads(), mc.cores = lx.options("mc.cores"),
  mc.preschedule = TRUE, mc.allow.recursive = FALSE)
}
\arguments{
\item{FUN}{the function to be applied. see \link{lapply}.}

\item{X}{an array, including a matrix. see \link{lapply}.}

\item{pg.verbose}{use progress bar}

\item{use.threads}{use multithreading}

\item{mc.cores}{The number of cores to use. see \link{mclapply}.}

\item{mc.preschedule}{perform prescheduling. see \link{mclapply}.}

\item{mc.allow.recursive}{allow recursive call. see \link{mclapply}.}
}
\description{
just like \link{Filter} but using \link{lx.lapply}
internally.
}
\examples{
lx.options(pg.verbose=TRUE)
lx.Filter(function(x) x\%\%2==0, 1:10)
lx.options(pg.verbose=FALSE)
lx.Filter(function(x) x\%\%2==0, 1:10)

}
