% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.findInterval}
\alias{lx.findInterval}
\title{open ends version of findInterval}
\usage{
lx.findInterval(x, vec)
}
\arguments{
\item{x}{numeric vector: the values to find}

\item{vec}{numeric, sorted (weakly) increasingly, of length N.
see \link{findInterval}}
}
\value{
integer vector of length length(x) with values in 1:N (and NA).
}
\description{
works as \link{findInterval} but always consider that end
breakpoints are infinite (i.e vec[1]=-Inf and vec[N]=+Inf).
This always return indices in range [1, N].
}
\note{
this is simply implemented as \code{pmax(1, findInterval(x, vec))}
}
\examples{
lx.findInterval(0:11, 1:10)
lx.findInterval(c(-Inf, Inf), 1:10)

}
\seealso{
\link{findInterval}
}
