% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.backports.r
\name{lx.get0}
\alias{lx.get0}
\title{backport of get0 for R < 3.2.0}
\usage{
lx.get0(x, envir = pos.to.env(-1L), mode = "any", inherits = TRUE,
  ifnotfound = NULL)
}
\arguments{
\item{x}{a variable name (given as a character string).}

\item{envir}{an environment to look in.}

\item{mode}{the mode or type of object sought.}

\item{inherits}{should the enclosing frames of the environment be searched?}

\item{ifnotfound}{the return value when x does not exist.}
}
\value{
The object - as from \link{get}(x, *) - if \link{exists}(x, *)
is true, otherwise \code{ifnotfound}
}
\description{
Look for an R object of the given name and possibly return it
}
\note{
code from \url{https://github.com/cran/backports/blob/master/R/get0.R}
}
\seealso{
\link{get0}
}
