% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.osutil.r
\name{lx.getargs}
\alias{lx.getargs}
\title{commandArgs parsing}
\usage{
lx.getargs(verbose = FALSE, defaults = NULL)
}
\arguments{
\item{verbose}{print verbage to screen}

\item{defaults}{a named list of defaults, optional}
}
\value{
a named list
}
\description{
return a named list of command line arguments

Usage:
call the R script as\cr
  ./myfile.R --args myarg=something\cr
or\cr
  R CMD BATCH --args myarg=something myfile.R\cr

Then in R do\cr
  myargs <- getArgs()\cr
to retrieve a named list of arguments
}
\note{
don't confuse this function with \link{lx.args} that handle function arguments.

Borrowed from : http://cwcode.wordpress.com/2013/04/16/the-joys-of-rscript/
}
\author{
Chris Wallace
}
