% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.gregexpr}
\alias{lx.gregexpr}
\title{gregexpr wrapper}
\usage{
lx.gregexpr(pattern, text, ...)
}
\arguments{
\item{pattern}{character string containing a regular expression}

\item{text}{character string where matches are sought}

\item{...}{any parameter of \link{gregexpr}}
}
\value{
integer vector of positions where pattern is found or empty vector if no match
}
\description{
similar to \link{gregexpr} but argument \code{text} is a single character
string and function returns a (possibly empty) integer vector of match positions.
}
\note{
the order of parameters is the same as in \link{gregexpr}, i.e. pattern first
}
\examples{
lx.gregexpr("i", "mississipi")
lx.gregexpr("[imsp]", "mississipi")
lx.gregexpr("o", "mississipi")

}
\seealso{
\code{\link{lx.strchr}}
}
