% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.head}
\alias{lx.head}
\title{head of vector}
\usage{
lx.head(x, n = 1L)
}
\arguments{
\item{x}{vector}

\item{n}{if positive take n first elements, if negative take all but n last elements}
}
\value{
vector of size n if n >= 0, or of size \code{length(x) - n} if n < 0
}
\description{
same as \link{head} but limited to vectors and quicker
}
\examples{
lx.head(1:10, 2)  # => 1 2
lx.head(1:10, -2) # => 1 2 3 4 5 6 7
}
\seealso{
\link{head} \link{lx.tail}
}
