% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.hist}
\alias{lx.hist}
\title{histogram plot with density}
\usage{
lx.hist(x, bw = "nrd0", freq = FALSE, main = paste("Histogram of", xname),
  xlab = xname, ..., col.density = "blue")
}
\arguments{
\item{x}{vector of values for which the histogram is desired}

\item{bw}{smoothing bandwidth (see \link{density})}

\item{freq}{y-axis is frequency or probability (see \link{hist})}

\item{main}{title (see \link{hist})}

\item{xlab}{x axis label (see \link{hist})}

\item{...}{other arguments to \link{hist}}

\item{col.density}{density line color}
}
\value{
same as \link{hist} plus a 'dns' component containing
the density.
}
\description{
histogram plot with density
}
\note{
\code{freq} defaults to FALSE
}
\examples{
\dontrun{
lx.hist(rnorm(1000))
lx.hist(rnorm(1000), main="Normal", xlab="x")
}
}
