% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.in}
\alias{lx.in}
\title{quicker \code{\%in\%} for numeric vectors}
\usage{
lx.in(x, table, .threshold = 500)
}
\arguments{
\item{x}{numeric vector: the values to be matched}

\item{table}{numeric vector: the values to be matched against.
(better if sorted increasingly)}

\item{.threshold}{internal threshold on table length to switch
between standard \%in\% and quicker algorithm.}
}
\value{
A logical vector, indicating if a match was located for each element of x.
}
\description{
works as \link{\%in\%} but much quicker and for numeric vectors only.
}
\note{
table should be sorted increasingly. if not it will be sorted
internally.
}
\examples{
lx.in(1:10, 3:5)
lx.in((1:10)/1000, (3:5)/1000)
lx.in(5e6, 1:1e7)  # quicker than 5e6 \%in\% 1:1e7

}
