% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.info}
\alias{lx.info}
\alias{lx.info}
\alias{lx.info<-}
\title{set/get information on object}
\usage{
lx.info(obj)

lx.info(obj) <- value
}
\arguments{
\item{obj}{an R object}

\item{value}{option value}
}
\description{
to get object information:\cr
\code{lx.info(obj)}\cr
to set object information:\cr
\code{lx.info(obj) <- "value"}\cr
}
\examples{
x <- 1:3
lx.info(x) <- "my array"
lx.info(x)
lx.doc(x)

}
\seealso{
\link{lx.doc}
}
