% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.globals.r
\docType{data}
\name{lx.iris}
\alias{lx.iris}
\title{nominal iris dataframe}
\description{
this is a version of \link{iris} with 3 nominal columns:\cr
species=c("setosa", "versicolor", "virginica")\cr
petal=c("small", "large")\cr
sepal=c("small", "large")\cr
used for testing and examples.
}
