% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.key.trans}
\alias{lx.key.trans}
\title{translate keys fom dictionnary}
\usage{
lx.key.trans(x, dict)
}
\arguments{
\item{x}{key or vector of keys}

\item{dict}{dictionnary (list indexed by keys, plus optional default value)}
}
\description{
translate keys in vector \code{x} to values from dictionnary \code{dict}
}
\examples{
lx.key.trans("red", c(red="rouge", blue="bleu"))
lx.key.trans(c("red", "blue"), c(red=1, green=2, blue=3))
lx.key.trans(c("red", "pink"), c(red=1, green=2, blue=3))
lx.key.trans(c("red", "pink"), c(red=1, green=2, blue=3, 0))
}
