% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.apply.r
\name{lx.lapply}
\alias{lx.lapply}
\title{lapply wrapper}
\usage{
lx.lapply(X, FUN, ..., pg.verbose = lx.options("pg.verbose"),
  use.threads = lx.use.threads(), mc.cores = lx.options("mc.cores"),
  mc.preschedule = TRUE, mc.allow.recursive = FALSE)
}
\arguments{
\item{X}{an array, including a matrix. see \link{lapply}.}

\item{FUN}{the function to be applied. see \link{lapply}.}

\item{...}{anything passed to FUN. see \link{lapply}.}

\item{pg.verbose}{use progress bar}

\item{use.threads}{use multithreading}

\item{mc.cores}{The number of cores to use. see \link{mclapply}.}

\item{mc.preschedule}{perform prescheduling. see \link{mclapply}.}

\item{mc.allow.recursive}{allow recursive call. see \link{mclapply}.}
}
\description{
\code{lx.lapply} will call different \link{lapply} flavors
depending upon the \code{use.threads} and \code{pg.verbose}.
arguments:\cr
if use.threads is TRUE then\cr
__use \link{mclapply}\cr
else if pg.verbose is TRUE then\cr
__use \link{lapply} with progress bar\cr
else\cr
__use \link{lapply} without progress bar\cr
}
\note{
default value for \code{use.threads} and \code{pg.verbose} are 
taken from \link{lx.options}
and may therefore be controlled globally.
}
\examples{
n <- 100
lx.options(pg.verbose=TRUE)
system.time(x <- lx.lapply(1:n, function(x) rnorm(5000*x)))
lx.options(pg.verbose=FALSE)
system.time(x <- lx.lapply(1:n, function(x) rnorm(5000*x)))

}
\seealso{
\link{lx.happly} for a version dulpicating file handle.
}
