% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.lazy}
\alias{lx.lazy}
\title{conditionaly restore or set object}
\usage{
lx.lazy(object, setfun, ..., force = FALSE)
}
\arguments{
\item{object}{to set (existing or not)}

\item{setfun}{function to set object}

\item{...}{arguments of setfun}

\item{force}{force reexecution of setfun even if dump file does exist}
}
\description{
conditionaly restore or set object depending upon
previously saved value
}
\details{
if file <object>.RData does exist then restore value from this file
(unless force=TRUE) else execute setfun(...) and save result into dump
file for later retrieval.
}
\examples{
foo <- function(n) sum(rnorm(n))  # dummy 
n <- 100000000
system.time(lx.lazy(mynorm, foo, n)) # 10 s
system.time(lx.lazy(mynorm, foo, n)) # <1 s
  
}
