% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.list2str}
\alias{lx.list2str}
\title{convert named list to character string}
\usage{
lx.list2str(lst, sep = "=", collapse = ", ", quote = NA, ...)
}
\arguments{
\item{lst}{named list}

\item{sep}{separator between key and value}

\item{collapse}{sep separator elements}

\item{quote}{character to use for quoting character values(use NA to prevent quoting)}

\item{...}{arguments to \link{format} for formating elements}
}
\value{
character string
}
\description{
convert named list to character string of the form
"key1<sep>value1<collapse>key2<sep>value2...."
}
\note{
list values should be atomic. vectors are printed as '[x,y,...]'.
      recursive lists or matrices are not (yet) handled.
}
\examples{
ll <- list(a=1, b="you", c=1:3)
lx.list2str(ll)
}
