% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.apply.r
\name{lx.Map}
\alias{lx.Map}
\title{Map wrapper}
\usage{
lx.Map(FUN, ..., use.threads = lx.use.threads())
}
\arguments{
\item{FUN}{the function to be applied. see \link{mapply}.}

\item{...}{arguments to vectorize over. see \link{mapply}.}

\item{use.threads}{use multithreading}
}
\description{
just like \link{Map} but using \link{lx.mapply}
internally\cr
}
