% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.numerical.r
\name{lx.maxima}
\alias{lx.maxima}
\title{get maxima of a curve}
\usage{
lx.maxima(x, span = 0.2, eps = 0.1, ...)
}
\arguments{
\item{x}{vector of numerical values}

\item{span}{pre-smoothing parameter (see details)}

\item{eps}{min ratio between highest and other reported maxima (see details)}

\item{...}{other parameters to \link{loess}}
}
\value{
vector of indices of maxima (ordered by decreasing heights)
}
\description{
get maxima of a curve
}
\details{
if \code{span > 0} then \code{x} is first smoothed using
\link{loess} with parameter \code{span}.
Then maxima are searched for. If \code{eps >= 0} then \code{x}
is expected to be a vector of positive values (typically a distribution) and only maxima
with height >= eps * max(height) are returned. If \code{eps < 0} then all maxima
are returned (\code{x} can be any curve with positive and negative values)
}
\note{
the technique used to locate extrema is quite simple (change of sign
of the derivative) and can easily be fooled by noisy data. this is the
reason why data is first smoothed. this function displays better behavior
on regular distributions provided by \link{lx.density}.
}
\examples{
x <- lx.density(c(rnorm(100, -1, 0.5), rnorm(100, 1, 0.5)))
x$x[lx.maxima(x$y)]
}
\seealso{
\link{lx.peaks}
}
