% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.mfrow}
\alias{lx.mfrow}
\title{determine best \code{\link{par}(mfrow=c(x,y))} layout to fit for n plots}
\usage{
lx.mfrow(n)
}
\arguments{
\item{n}{number of plots}
}
\description{
determine best \code{\link{par}(mfrow=c(x,y))} layout to fit for n plots
}
\examples{
opar <- par(no.readonly=TRUE)
par(mfrow=lx.mfrow(10))
# ... plots
par(opar)

}
\seealso{
\link{par}
}
