% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.new}
\alias{lx.new}
\title{create an S3 instance of given class}
\usage{
lx.new(classname, ...)
}
\arguments{
\item{classname}{string or string vector of classnames}

\item{...}{anything accepted by \link{list}}
}
\value{
a named list of \code{...} named fields and class(es) classname
}
\description{
create an S3 instance of given class
}
\examples{
x <- lx.new("people", name="alain", age=10)
class(x)
x$age

}
\seealso{
\link{class}
}
