% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.open}
\alias{lx.open}
\title{open (binary) file}
\usage{
lx.open(filename, mode = "rb", .handler = NULL)
}
\arguments{
\item{filename}{file name}

\item{mode}{open mode (see \link{HELP.FILE.HANDLE})}

\item{.handler}{force use of given handler name (internal use only)}
}
\value{
file handle
}
\description{
open (binary) file
}
\note{
if you force handler (thru .handler) the handler$accept test function
will not be called.
}
\seealso{
\link{HELP.FILE.HANDLE}
}
