% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.numerical.r
\name{lx.peaks}
\alias{lx.peaks}
\title{get peaks of a curve}
\usage{
lx.peaks(x, span = 0.2, eps.x = 0.1, eps.dx = eps.x, na.val = NA, ...)
}
\arguments{
\item{x}{vector of numerical values (typically a distribution)}

\item{span}{pre-smoothing parameter (see details)}

\item{eps.x}{eps parameter for x maxima (see details)}

\item{eps.dx}{eps parameter for x derivative maxima (see details)}

\item{na.val}{default width (see details)}

\item{...}{other parameters to \link{loess}}
}
\value{
list of three components:\cr
\preformatted{
 - pos   :  vector of indices of peaks position (ordered by decreasing heights)
 - left  :  vector of indices of peaks left side
 - right :  vector of indices of peaks right side
}
}
\description{
get peaks position and width of a curve
}
\details{
if \code{span > 0} then \code{x} is first smoothed using
\link{loess} with parameter \code{span}.
Then, maxima of x are search with \link{lx.maxima} using eps.x parameter.
Then, foreach maximum, the closest maxima of x derivative are searched
for to compute left and right peak boundaries. if no boundary is found
then value is set to maximum.pos +/- \code{na.val}.
}
\note{
if x is a normal distribution then band width is an estimate of 2*sd

the technique used to locate extrema is quite simple (change of sign
of the derivative) and can easily be fooled by noisy data. this is the
reason why data is first smoothed. this function displays better behavior
on regular distributions provided by \link{lx.density}.
}
\examples{
x <- lx.density(c(rnorm(100, -1, 0.3), rnorm(100, 1, 0.3)))
p <- lx.peaks(x$y)
\dontrun{
  plot(x)
  abline(v=x$x[p$pos], col=1)
  abline(v=x$x[c(p$left, p$right)], col=2)
}
}
\seealso{
\link{lx.maxima}
}
