% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.plot.inset}
\alias{lx.plot.inset}
\title{add inset plot into current plot}
\usage{
lx.plot.inset(..., inset.pos = c("tr", "tl", "br", "bl"), inset.alpha = 0.5,
  inset.margin = 0.01)
}
\arguments{
\item{...}{same arguments as in generic \link{plot}}

\item{inset.pos}{inset position 'tl', 'tr', 'bl', 'br'}

\item{inset.alpha}{inset size factor}

\item{inset.margin}{inset margin}
}
\description{
add a small (inset) plot in current plot
}
\details{
inset position indicate the poistion of the inset as tl:top-left,
tr:top-right, bl=bottom-left, br:bottom-right
}
\examples{
\dontrun{
plot(function(x) {sin(x*10)/x}, col="red", ylab="", main="main-plot")
lx.plot.inset(function(x) {cos(x*20)}, ylab="", xlab="inset", cex.axis=0.5,
              inset.pos="tr", inset.alpha=0.5)
}
}
