% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.plot.r
\name{lx.plot}
\alias{lx.plot}
\title{plotting large number of points}
\usage{
lx.plot(x, y = NULL, fun = plot, cex = 0.1, xlab = NULL, ylab = NULL,
  col = "black", pch = 1, ..., .samp = 10000, .method = c("sampled",
  "spaced"), .seed = 0)
}
\arguments{
\item{x, y}{data (same as in \link{plot})}

\item{fun}{plotting function (e.g. \link{plot}, \link{points}, \link{lines})}

\item{cex}{plot symbols magnifying factor. see \link{plot}}

\item{xlab}{a label for the x axis, defaults to a description of x}

\item{ylab}{a label for the y axis, defaults to a description of y}

\item{col}{plotting color (see \link{par})}

\item{pch}{plotting character (see \link{par})}

\item{...}{arguments to \code{fun} (see \link{graphical parameters})}

\item{.samp}{number of points to sample}

\item{.method}{sampling method (may be abbreviated)\cr
\code{sampled}: random sampling\cr
\code{spaced}: equally spaced points\cr}

\item{.seed}{seed for sampling (for .method="sampled" only). use
NA to prevent seeding (i.e. to use current seed). use a non-NA
value to produce reproducible behavior.}
}
\value{
same as \link{barplot}
}
\description{
this is a wrapper to \link{plot} (or \link{points}, \link{lines})
with pre-sampling of points. It is intended to plot very large
number of points (i.e. when plotting time becomes prohibited).
}
\note{
if \code{.seed != NA} the current seed will be saved and restored.
}
\examples{
\dontrun{
lx.plot(rnorm(1e7))
lx.plot(rnorm(1e7), ylab="Normal")
lx.plot(rnorm(1e7, 0, 0.1), fun=points, col=2)
}
}
