% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.prm.exp}
\alias{lx.prm.exp}
\alias{lx.dexp}
\alias{lx.pexp}
\alias{lx.rexp}
\title{Re-parameterized Exponential distribution}
\usage{
lx.prm.exp(mean = 1)

lx.dexp(x, mean = 1, ...)

lx.pexp(x, mean = 1, ...)

lx.rexp(n, mean = 1, ...)
}
\arguments{
\item{mean}{vector of means}

\item{x}{vector of quantiles}

\item{...}{other parameters to \link{dexp}, \link{pexp} and \link{rexp}}

\item{n}{number of observations}
}
\value{
\bold{lx.prm.exp} returns the standard parameters
of R library functions.
\bold{lx.dexp} gives the density, \bold{lx.pexp} gives the distribution function
and \bold{lx.rexp} generates random deviates.
}
\description{
Exponential  distibution reparameterized with mean
}
\section{Functions}{
\itemize{
\item \code{lx.dexp}: density function

\item \code{lx.pexp}: distribution function

\item \code{lx.rexp}: random deviate
}}

\seealso{
\link{dexp}, \link{pexp} and \link{rexp}
}
