% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.prm.gamma}
\alias{lx.prm.gamma}
\alias{lx.dgamma}
\alias{lx.pgamma}
\alias{lx.rgamma}
\title{Re-parameterized Gamma distribution}
\usage{
lx.prm.gamma(mean, sd)

lx.dgamma(x, mean, sd, ...)

lx.pgamma(x, mean, sd, ...)

lx.rgamma(n, mean, sd, ...)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{x}{vector of quantiles}

\item{...}{other parameters to \link{dgamma}, \link{pgamma} and \link{rgamma}}

\item{n}{number of observations}
}
\value{
\bold{lx.prm.gamma} returns the standard parameters
of R library functions.
\bold{lx.dgamma} gives the density, \bold{lx.pgamma} gives the distribution function
and \bold{lx.rgamma} generates random deviates.
}
\description{
Gamma distibution reparameterized with mean and variance
}
\section{Functions}{
\itemize{
\item \code{lx.dgamma}: density function

\item \code{lx.pgamma}: distribution function

\item \code{lx.rgamma}: random deviate
}}

\seealso{
\link{dgamma}, \link{pgamma} and \link{rgamma}
}
