% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.prm.nbinom}
\alias{lx.prm.nbinom}
\alias{lx.dnbinom}
\alias{lx.pnbinom}
\alias{lx.rnbinom}
\title{Re-parameterized Negative Binomial distribution}
\usage{
lx.prm.nbinom(mean, sd)

lx.dnbinom(x, mean, sd, ...)

lx.pnbinom(x, mean, sd, ...)

lx.rnbinom(n, mean, sd, ...)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{x}{vector of quantiles}

\item{...}{other parameters to \link{dnbinom}, \link{pnbinom} and \link{rnbinom}}

\item{n}{number of observations}
}
\value{
\bold{lx.prm.nbinom} returns the standard parameters
of R library functions.
\bold{lx.dnbinom} gives the density, \bold{lx.pnbinom} gives the distribution function
and \bold{lx.rnbinom} generates random deviates.
}
\description{
Negative Binomial  distibution reparameterized with mean and variance
}
\section{Functions}{
\itemize{
\item \code{lx.dnbinom}: density function

\item \code{lx.pnbinom}: distribution function

\item \code{lx.rnbinom}: random deviate
}}

\seealso{
\link{dnbinom}, \link{pnbinom} and \link{rnbinom}
}
