% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.prm.norm}
\alias{lx.prm.norm}
\alias{lx.dnorm}
\alias{lx.pnorm}
\alias{lx.rnorm}
\title{Normal distribution}
\usage{
lx.prm.norm(mean, sd)

lx.dnorm(x, mean = 0, sd = 1, ...)

lx.pnorm(x, mean = 0, sd = 1, ...)

lx.rnorm(n, mean = 0, sd = 1, ...)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{x}{vector of quantiles}

\item{...}{other parameters to \link{dnorm}, \link{pnorm} and \link{rnorm}}

\item{n}{number of observations}
}
\value{
\bold{lx.prm.norm} returns the standard parameters
of R library functions.
\bold{lx.dnorm} gives the density, \bold{lx.pnorm} gives the distribution function
and \bold{lx.rnorm} generates random deviates.
}
\description{
These functions are equivalent to \link{dnorm}, \link{pnorm} and \link{rnorm}.
they are just provided for symetry with other re-parametrized distributions.
}
\section{Functions}{
\itemize{
\item \code{lx.dnorm}: density function

\item \code{lx.pnorm}: distribution function

\item \code{lx.rnorm}: random deviate
}}

\seealso{
\link{dnorm}, \link{pnorm} and \link{rnorm}
}
