% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.prm.pois}
\alias{lx.prm.pois}
\alias{lx.dpois}
\alias{lx.ppois}
\alias{lx.rpois}
\title{Poisson distribution}
\usage{
lx.prm.pois(mean)

lx.dpois(x, mean, ...)

lx.ppois(x, mean, ...)

lx.rpois(n, mean, ...)
}
\arguments{
\item{mean}{vector of (non-negative) means}

\item{x}{vector of quantiles}

\item{...}{other parameters to \link{dpois}, \link{ppois} and \link{rpois}}

\item{n}{number of observations}
}
\value{
\bold{lx.prm.pois} returns the standard parameters
of R library functions.
\bold{lx.dpois} gives the density, \bold{lx.ppois} gives the distribution function
and \bold{lx.rpois} generates random deviates.
}
\description{
These functions are equivalent to \link{dpois}, \link{ppois} and \link{rpois}.
they are just provided for symetry with other re-parametrized distributions.
}
\section{Functions}{
\itemize{
\item \code{lx.dpois}: density function

\item \code{lx.ppois}: distribution function

\item \code{lx.rpois}: random deviate
}}

\seealso{
\link{dpois}, \link{ppois} and \link{rpois}
}
