% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.r
\docType{package}
\name{lx}
\alias{lx}
\alias{lx-package}
\title{LX utilities}
\description{
Generic utilities used in LX packages.\cr

These utilities are currently subdivised in different subpackages:\cr
\subsection{General programming utilities}{
functions are prefixed by 'lx'\cr
see \code{??lx} for more
}

\subsection{Latex/pdf reports}{
functions are prefixed by 'tex'\cr
see \code{??tex} for more
}
}
\details{
\tabular{ll}{
Package: \tab lx\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-22\cr
License: \tab GPL\cr
}
}
\author{
Alain Viari
}
