% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.read.int16}
\alias{lx.read.int16}
\title{read binary 16 bits integer(s)}
\usage{
lx.read.int16(handle, n = 1L, signed = FALSE, little.endian = TRUE)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}

\item{n}{number of integers to read}

\item{signed}{read signed integer(s)}

\item{little.endian}{integer encoding is little endian}
}
\value{
(vector of) integer(s)
}
\description{
read binary 16 bits integer(s)
}
\note{
n is an integer, therefore limited to 2,147,483,647
}
\examples{
fh <- lx.open(lx.system.file("samples/test_bin.bst"))
x <- lx.read.int16(fh, n=2)
lx.close(fh)
}
