% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.read.int32}
\alias{lx.read.int32}
\title{read binary 32 bits integer(s)}
\usage{
lx.read.int32(handle, n = 1L, signed = FALSE, little.endian = TRUE)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}

\item{n}{number of integers to read}

\item{signed}{read signed integer(s)}

\item{little.endian}{integer encoding is little endian}
}
\value{
(vector of) integer(s) if signed == TRUE
else (vector of) double(s) without loss of precision
}
\description{
read binary 32 bits integer(s)
}
\note{
n is an integer, therefore limited to 2,147,483,647
}
\examples{
fh <- lx.open(lx.system.file("samples/test_bin.bst"))
cod  <- lx.read.int32(fh, n=1)
nseq <- lx.read.int32(fh, n=1)
s1 <- lx.read.string(fh)
siz1 <- lx.read.int64(fh, n=1) # 24
crc1 <- lx.read.int32(fh, n=1)
s2 <- lx.read.string(fh)
siz2 <- lx.read.int64(fh, n=1) # 32
crc2 <- lx.read.int32(fh, n=1)
ss1 <- lx.read.char(fh, n=siz1) # "ACGTACGTACGTAAAAACGTACGT"
ss2 <- lx.read.char(fh, n=siz2) # "GCGCGCGCGCGCGCGCGCGCGCGCTTTTTATATATATAX"
lx.close(fh)

}
