% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.read.int8}
\alias{lx.read.int8}
\title{read binary 8 bits integer(s)}
\usage{
lx.read.int8(handle, n = 1L, signed = FALSE)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}

\item{n}{number of integers to read}

\item{signed}{read signed integer(s)}
}
\value{
(vector of) integer(s)
}
\description{
read binary 8 bits integer(s)
}
\note{
n is an integer, therefore limited to 2,147,483,647
}
\examples{
fh <- lx.open(lx.system.file("samples/test_bin.bst"))
x <- lx.read.int8(fh, n=4)
lx.close(fh)
}
