% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.read.string}
\alias{lx.read.string}
\title{read binary C-style or basta string}
\usage{
lx.read.string(handle, with.size = TRUE)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}

\item{with.size}{if true basta string else simple C-style string}
}
\value{
character string
}
\description{
a C-style string is null terminated character array\cr
a basta string is binary encoded in the following way\cr
\itemize{
 \item int32 : string size
 \item chars+NULL : null terminated (C style) character array
}
}
\note{
if with.size==FALSE this function is equivalent to
\code{lx.read.char(handle, n=-1L, skip=FALSE)}
}
\examples{
fh <- lx.open(lx.system.file("samples/test_bin.bst"))
cod  <- lx.read.int32(fh, n=1)
nseq <- lx.read.int32(fh, n=1)
s1 <- lx.read.string(fh)
siz1 <- lx.read.int64(fh, n=1)
crc1 <- lx.read.int32(fh, n=1)
s2 <- lx.read.string(fh)
siz2 <- lx.read.int64(fh, n=1)
crc2 <- lx.read.int32(fh, n=1)
ss1 <- lx.read.char(fh, n=siz1)
ss2 <- lx.read.char(fh, n=siz2)
lx.close(fh)
}
