% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.recycle}
\alias{lx.recycle}
\title{expand by recycling or shrink vector to given size}
\usage{
lx.recycle(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{positive integer}
}
\value{
vector of size n
}
\description{
given a vector x either shrink or expand by recycling to size n
}
\examples{
lx.recycle(1:3, 6)
lx.recycle(1:10, 6)
lx.recycle(1:10, 0)
lx.recycle(NULL, 2)
}
\seealso{
\link{lx.rotate}
}
