% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.regex.quote}
\alias{lx.regex.quote}
\title{quote special characters used in regex}
\usage{
lx.regex.quote(s)
}
\arguments{
\item{s}{string or string vector}
}
\description{
quote special characters used in regex
}
\examples{
lx.regex.quote("the (lazy) dog")
}
