% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.register.file.handler}
\alias{lx.register.file.handler}
\title{register handler}
\usage{
lx.register.file.handler(handler)
}
\arguments{
\item{handler}{: file handler}
}
\description{
register handler at topmost position in handlers stack
}
\note{
if a handler with same name already exists in stack it is first
      removed and the new handler is added topmost.
}
