% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.remove.file.handler}
\alias{lx.remove.file.handler}
\title{remove file handler(s)}
\usage{
lx.remove.file.handler(name)
}
\arguments{
\item{name}{: name(s) of handler(s) to remove}
}
\description{
remove handler from handlers stack
}
