% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.restore}
\alias{lx.restore}
\title{restore objects from dump file}
\usage{
lx.restore(name)
}
\arguments{
\item{name}{of first object used in \code{lx.save}}
}
\description{
restore objects from dump file
}
\examples{
data(lx.iris)
tab1 <- tab2 <- lx.iris
lx.save(tab1, tab2)
remove(tab1, tab2)
lx.restore("tab1")
}
\seealso{
\code{\link{lx.save}}
}
