% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.rev.dict}
\alias{lx.rev.dict}
\title{reverse dictionnary}
\usage{
lx.rev.dict(dict)
}
\arguments{
\item{dict}{dictionnary : a list indexed by keys
(plus optional default value that is ignored)}
}
\value{
reverse dictionary : a list indexed by values
}
\description{
reverse dictionnary
}
\examples{
dict <- list(warm=c("red"), cold=c("blue", "green"), basic=c("red", "blue", "green"), "white")
lx.rev.dict(dict)
# => list(red=c("warm", "basic"), blue=c("cold", "basic"), green=c("cold", "basic"))
}
