% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.rewind}
\alias{lx.rewind}
\title{rewind file handle}
\usage{
lx.rewind(handle)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}
}
\value{
handle (invisible)
}
\description{
rewind file handle opened by \link{lx.open}
}
\note{
this is a shorthand for \code{lx.seek(handle, 0)}
}
