% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.rnd.pop}
\alias{lx.rnd.pop}
\title{restore current global random seed}
\usage{
lx.rnd.pop()
}
\value{
(invisible) restored random seed
}
\description{
pop current global random seed (.Random.seed) from internal stack
this has to be used (in pairs with \link{lx.rnd.push}) to
save current random seed and restore it later
}
\examples{
set.seed(1)
x <- sample(letters, 3)
set.seed(1)
lx.rnd.push() # same as
set.seed(2)   # lx.rnd.push(2)
y <- sample(letters, 3)
lx.rnd.pop()
z <- sample(letters, 3)
identical(x, z)
}
