% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.rnd.push}
\alias{lx.rnd.push}
\title{save current global random seed}
\usage{
lx.rnd.push(seed)
}
\arguments{
\item{seed}{\bold{optional} parameter, if specified then additionnaly
\link{set.seed}(seed)}
}
\value{
(invisible) current random seed
}
\description{
push current global random seed (.Random.seed) into internal stack
this has to be used (in pairs with \link{lx.rnd.pop}) to
save current random seed and restore it later
}
\examples{
set.seed(1)
x <- sample(letters, 3)
set.seed(1)
lx.rnd.push() # same as
set.seed(2)   # lx.rnd.push(2)
y <- sample(letters, 3)
lx.rnd.pop()
z <- sample(letters, 3)
identical(x, z)
}
