% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.rowMins}
\alias{lx.rowMins}
\title{get minimum in rows}
\usage{
lx.rowMins(x, ties.method = c("random", "first", "last"), what = c("value",
  "index"))
}
\arguments{
\item{x}{numerical matrix}

\item{ties.method}{a character string specifying how ties are handled. (see \link{max.col})}

\item{what}{a character string specifying what should be returned (see Value)}
}
\value{
\itemize{
\item if \code{what == "value"} return a minimum value for each row
\item if \code{what == "index"} return index of a minimum value for each row
}
}
\description{
find the minimum for each row of a matrix
}
\note{
with \code{what == "index"} this function is identical
to \link{max.col}(-x)

with \code{what == "value"} this function is identical to 
\code{apply(1, x, min)} but much quicker for large matrices
}
\seealso{
\link{lx.rowMaxs}
}
