% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.sample}
\alias{lx.sample}
\title{base::sample wrapper}
\usage{
lx.sample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose,
or a positive integer. (see \link{sample})}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement}

\item{prob}{a vector of probability weights for obtaining the elements
of the vector being sampled.}
}
\value{
a vector of length size with elements drawn from either x
or from the integers 1:x
}
\description{
this function calls \code{base::sample(x, size, prob, replace)} except
for the case where \code{prob != NULL and replace == FALSE}. In this
case \code{base::sample} is very slow (actually quadratic in the size of
\code{x} and of \code{size}. This is replaced by the 'reservoir' technique
of Efraimidis & Spirakis (see note)
}
\note{
\bold{reference}: P. Efraimidis and P.Spirakis, Information Processing Letters,
97, 181-185 (2006).
}
