% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.apply.r
\name{lx.sapply}
\alias{lx.sapply}
\title{sapply wrapper}
\usage{
lx.sapply(X, FUN, ..., pg.verbose = lx.options("pg.verbose"))
}
\arguments{
\item{X}{an array, including a matrix. see \link{lapply}.}

\item{FUN}{the function to be applied. see \link{lapply}.}

\item{...}{anything passed to FUN. see \link{lapply}.}

\item{pg.verbose}{use progress bar}
}
\description{
\code{lx.sapply} will call different \link{sapply} flavors
depending upon the \code{pg.verbose}.argument:\cr
if pg.verbose is TRUE then\cr
__use \link{sapply} with progress bar\cr
else\cr
__use \link{sapply} without progress bar\cr
}
\note{
default value for  \code{pg.verbose} is taken from \link{lx.options}
and may therefore be controlled globally.

there is no multithread flavor yet, see \link{lx.lapply}
}
\examples{
n <- 100
lx.options(pg.verbose=TRUE)
system.time(x <- lx.sapply(1:n, function(x) rnorm(5000*x)))
lx.options(pg.verbose=FALSE)
system.time(x <- lx.sapply(1:n, function(x) rnorm(5000*x)))

}
