% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.save}
\alias{lx.save}
\title{save objects into dump file}
\usage{
lx.save(...)
}
\arguments{
\item{...}{list of objects to save}
}
\description{
save objects into dump file
}
\details{
the dump file name is <obj1>.RData
         where obj1 is the first object in \code{...}
}
\examples{
data(lx.iris)
tab1 <- tab2 <- lx.iris
lx.save(tab1, tab2)
remove(tab1, tab2)
lx.restore("tab1")
}
\seealso{
\code{\link{lx.restore}}
}
