% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.saved}
\alias{lx.saved}
\title{tells if object has been previously saved}
\usage{
lx.saved(name)
}
\arguments{
\item{name}{of first object used in \code{lx.save}}
}
\description{
tells if object has been previously saved
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.save(tab)
lx.saved("tab")
}
\seealso{
\code{\link{lx.save}}
}
