% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.scale}
\alias{lx.scale}
\title{linearly rescale numeric vector to specified range.}
\usage{
lx.scale(x, to = c(0, 1), from = range(x, na.rm = TRUE), zerob = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{to}{output range (numeric vector of length two)}

\item{from}{input range (numeric vector of length two).\cr
if not specified, use the range of x}

\item{zerob}{force b=0. this is useful for rescaling differences}
}
\value{
rescaled vector
}
\description{
rescale numeric vector x to x' from range [f1, f2] to range [t1, t2]
according to:\cr
\code{x}'\code{ = a.x + b}\cr
where a=(t2-t1)/(f2-f1) and b = t1 - a.f1
}
\examples{
lx.scale(1:10)
lx.scale(1:10, c(1,10))
}
