% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{lx.seek}
\alias{lx.seek}
\title{tell/seek file}
\usage{
lx.seek(handle, ...)
}
\arguments{
\item{handle}{file handle (opened by \link{lx.open})}

\item{...}{optional position in file}
}
\value{
current position (before any move) as a (numeric) byte offset from the origin.
}
\description{
reposition file handle opened by \link{lx.open}
}
\details{
if ... is absent then returns current position (ftell) else
set the position to (numeric) argument.
}
\note{
for default (R) handler, position may exceed 32 bit integer.
}
\examples{
fh <- lx.open(lx.system.file("samples/test_bin.bst"))
cod  <- lx.read.int32(fh, n=1)
pos <- lx.seek(fh)
nseq <- lx.read.int32(fh, n=1)
lx.seek(fh, 0)
cod  <- lx.read.int32(fh, n=1)
nseq <- lx.read.int32(fh, n=1)
lx.seek(fh, pos)
nseq <- lx.read.int32(fh, n=1)
lx.close(fh)

}
