% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.serialize}
\alias{lx.serialize}
\title{serialize single object into file}
\usage{
lx.serialize(object, name = NULL, ...)
}
\arguments{
\item{object}{object to serialize}

\item{name}{file name (without extension)
if NULL then use symbol object}

\item{...}{additionnal options to \code{\link{saveRDS}}}
}
\description{
serialize single object into file
}
\details{
object is serialized into file named \code{name.rds}
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.serialize(tab)  # create file \\code{tab.rds}
duptab <- lx.unserialize("tab")
}
\seealso{
\code{\link{lx.unserialize}}
}
