% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.serialized}
\alias{lx.serialized}
\title{tells if serialized object does exist}
\usage{
lx.serialized(name)
}
\arguments{
\item{name}{of object used in \code{lx.serialized}}
}
\description{
tells if serialized object does exist
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.serialize(tab)  # create file tab.Rdata
lx.serialized("tab")
}
\seealso{
\code{\link{lx.serialize}}
}
