% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.shift}
\alias{lx.shift}
\title{shift vector n times and pad with fill value}
\usage{
lx.shift(x, n = 1L, fill = NA)
}
\arguments{
\item{x}{vector}

\item{n}{number of shifts.\cr
if >= 0 shift to left else shift to right}

\item{fill}{value to fill}
}
\value{
shifted vector
}
\description{
shift vector n times and pad with fill value
}
\examples{
lx.shift(1:5, 2, fill=0)  # => 3 4 5 0 0
lx.shift(1:5, -2, fill=0) # => 0 0 1 2 3
}
