% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.numerical.r
\name{lx.smooth.median}
\alias{lx.smooth.median}
\title{Tukey's median smoothing}
\usage{
lx.smooth.median(x, k = 3L, tol = 0)
}
\arguments{
\item{x}{numeric vector, the \emph{dependent} variable to be smoothed}

\item{k}{vector of successive smoothing windows}

\item{tol}{tolerance for smoothing iterations (see details)}
}
\value{
vector of smoothed values of the same length as x
}
\description{
smooth data using iterative Tukey's median smoothing. 
this is typically intented for piecewise constant data (step function)
to smooth signal whuile preserving edges.
see \code{http://dx.doi.org/10.1007/BFb0057597}
}
\details{
the procedure runs as follows:\cr
\preformatted{
foreach window.size in k
  repeat
    x0 <- x
    x  <- runmed(x0, window.size)
  while (rss(x0,x) < tol)
end
}
set \code{tol} to Inf to prevent iterations and to 0 (default)
for full iterations.
}
\examples{
x <- c(rnorm(100, -1, 0.1), rnorm(100, 1, 0.1))
sx <- lx.smooth.median(x, c(3, 5, 15, 35))
\dontrun{
plot(x)
lines(sx, col=2)
}
}
\seealso{
\link{lx.loess}, \link{runmed}
}
