% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.stack.r
\name{lx.stack.is.empty}
\alias{lx.stack.is.empty}
\title{tell if stack is empty}
\usage{
lx.stack.is.empty(stk)
}
\arguments{
\item{stk}{Stack}
}
\value{
TRUE if stack is empty
}
\description{
tell if stack is empty
}
\note{
this implementation uses lists and is not efficient.
this is mostly intended for storing small number of elements.
if you need a better implementation, use the 'rstack' library
}
\examples{
stk <- lx.stack.new()
stk <- lx.stack.push(stk, 1)
stk <- lx.stack.push(stk, "two")
stk <- lx.stack.push(stk, NULL)
stk <- lx.stack.push(stk, 1:4)
while (! lx.stack.is.empty(stk)) {
 stk <- lx.stack.pop(stk)
 val <- lx.stack.value(stk)
 lx.out(typeof(val), ": ", val)
}
}
