% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.logging.r
\name{lx.stopif}
\alias{lx.stopif}
\title{stop execution if condition is true}
\usage{
lx.stopif(condition, ..., level = "error", up.frame = 0L, trace = TRUE)
}
\arguments{
\item{condition}{boolean value}

\item{...}{anything accepted by \link{lx.out}}

\item{level}{see \link{lx.out}}

\item{up.frame}{see \link{lx.out}}

\item{trace}{printout traceback}
}
\value{
invisible(condition)
}
\description{
stop execution if condition is true
}
\examples{
\dontrun{
x <- -1
lx.stopif(x < 0, "x=", x, " is negative")
}

}
\seealso{
\link{lx.out}, \link{lx.warn}, \link{lx.warnif}
}
