% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.strchr}
\alias{lx.strchr}
\title{find occurences of character(s) in string}
\usage{
lx.strchr(s, chr = "")
}
\arguments{
\item{s}{string where matches are sought}

\item{chr}{string containing any character to find}
}
\value{
integer vector of positions where any char from chr is found or empty vector if no match
}
\description{
find occurences of character(s) in string
}
\examples{
lx.strchr("mississipi", "i")
lx.strchr("mississipi", "imsp")
lx.strchr("mississipi", "o")
}
\seealso{
\code{\link{lx.gregexpr}}
}
