% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.strrev}
\alias{lx.strrev}
\title{reverse string}
\usage{
lx.strrev(s)
}
\arguments{
\item{s}{string or string vector}
}
\value{
vector of reversed strings
}
\description{
reverse string
}
\examples{
lx.strrev("the lazy dog")
}
