% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.strsplit}
\alias{lx.strsplit}
\title{split string around regexp}
\usage{
lx.strsplit(s, split = "( |\\t)+", trim = TRUE, ...)
}
\arguments{
\item{s}{string (or string vector)}

\item{split}{string (or string vector) containing regular expression(s)
(unless you specify \code{fixed = TRUE} in ...) to use for splitting.
see \link{strsplit}}

\item{trim}{also trim results using \link{lx.strtrim}}

\item{...}{other argument to \link{strsplit}}
}
\value{
vector of strings (or list of vectors of strings, see Details)
}
\description{
this is awrapper around \link{strsplit} to simplify
call when both s and split are simple strings.
}
\details{
the type of result depends upon the length of s. if s is a vector (length > 1)
then this function behave exactly as \link{strsplit} and returns a list of
string vectors. if s is a simple string (as this is often the case) then this
function returns a simple string vector.
}
\examples{
lx.strsplit("the lazy dog")
# is same as
strsplit("the lazy dog", " ")[[1]]
lx.strsplit(c("the lazy", "black dog"))
lx.strsplit("the lazy dog", "y")
lx.strsplit("the lazy dog", "y", fixed=TRUE)
}
