% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.strings.r
\name{lx.strtrim}
\alias{lx.strtrim}
\title{trim leading or trailing blanks (spaces or tabs)
in string}
\usage{
lx.strtrim(s, side = c("both", "left", "right"))
}
\arguments{
\item{s}{string or string vector}

\item{side}{c('left', 'right', 'both'). which end to trim}
}
\description{
trim leading or trailing blanks (spaces or tabs)
in string
}
\examples{
paste(">", lx.strtrim(" the lazy dog  "), "<", sep="")
}
