% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.summary}
\alias{lx.summary}
\title{a prettier \code{summary} for dataframes}
\usage{
lx.summary(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
a prettier \code{summary} for dataframes
}
\note{
only works with numeric columns
}
\examples{
lx.summary(iris[,-5])
}
