% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.osutil.r
\name{lx.sysinfo}
\alias{lx.sysinfo}
\title{wrapper to \link{Sys.info}}
\usage{
lx.sysinfo(...)
}
\arguments{
\item{...}{0 or more keys to \link{Sys.info} (if empty, then
return all entries)}
}
\value{
a (possibly empty) vector of entries of \link{Sys.info}
}
\description{
get system and user information
}
\note{
result may be NULL if \link{Sys.info} is not implemented
}
\examples{
x <- lx.sysinfo()
x <- lx.sysinfo("user", "machine")
}
