% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.packages.r
\name{lx.system.file}
\alias{lx.system.file}
\title{find names of R system files}
\usage{
lx.system.file(filename = "", package = "lx")
}
\arguments{
\item{filename}{filename to append to path}

\item{package}{package name}
}
\value{
<packageRootDir>/[filename]
}
\description{
this is nearly identical to \code{system.file}
but does not check at all for file existence.
}
