% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.bycols}
\alias{lx.table.bycols}
\title{k-dimensional contingency table(s) from columns}
\usage{
lx.table.bycols(tab, by = colnames(tab))
}
\arguments{
\item{tab}{a table (matrix or dataframe) of nominal values}

\item{by}{a list of column indices or names (of length k >= 2)}
}
\value{
k-dimensional contingency table, defined as the
        contigency table of the first two columns stratified on other columns
}
\description{
k-dimensional contingency table(s) from columns
}
\note{
this is the same as \code{table(tab[,by])}
}
\examples{
data(lx.iris)
tab <- lx.iris
# default: stratify by last dimensions (here sepal)
lx.table.bycols(tab)
# stratify by species (column index 1)
lx.table.bycols(tab, c(2,3,1))
# same as
lx.table.bycols(tab, c("petal", "sepal", "species"))
}
\seealso{
\link{lx.tables}
}
