% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.bycoocs}
\alias{lx.table.bycoocs}
\title{cross table of co-occurences in sets}
\usage{
lx.table.bycoocs(lsets, levels = NULL)
}
\arguments{
\item{lsets}{a list of sets Si}

\item{levels}{elements to keep (if NULL keeps all elements)}
}
\value{
n x n table where n = |union(Si)| and entry T[i,j] equals the number
        of co-occurences of elements i and j in sets.
        diagonal elements are the number of occurences of each element.
}
\description{
cross table of co-occurences in sets
}
\examples{
l <- list(c('a'), c('a', 'b'), c('a', 'b', 'c'))
lx.table.bycoocs(l)

}
\seealso{
\link{lx.tables}
}
