% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.byfacts}
\alias{lx.table.byfacts}
\title{make contingency table(s) by factors in columns}
\usage{
lx.table.byfacts(tab, by = colnames(tab))
}
\arguments{
\item{tab}{a table (matrix or dataframe) of nominal values}

\item{by}{a list of column indices or names (of length k >= 1)}
}
\value{
a 2-dimensional contingency table T of size f x k,
        where f = number of levels in \code{tab[,by]}
        and T[i,j] equals to the number of rows with factor i in column j
}
\description{
make contingency table(s) by factors in columns
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.table.byfacts(tab)
}
\seealso{
\link{lx.tables}
}
