% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.bymsets}
\alias{lx.table.bymsets}
\title{cross table of multisets intersections}
\usage{
lx.table.bymsets(lsets, as.card = FALSE)
}
\arguments{
\item{lsets}{a list of n named multisets Mi
each multiset can be specified as a list with repeated elements
e.g. c('a', 'a', 'b', 'c', 'b')
or, with as.card=TRUE, as a numeric list 
e.g. c(a=2, b=2, c=1)}

\item{as.card}{use cardinality in multiset definition}
}
\value{
n x m table T, where m = number of different elements in all Mi
        and T [i, j] equals to the number of elements j in multiset i
}
\description{
cross table of multisets intersections
}
\examples{
lsets <- list(A=c('a', 'a', 'b'), B=c('a'), C=c('a', 'b', 'b', 'c'), D=c('c'))
lx.table.bymsets(lsets)
lsets <- list(A=c(a=2, b=1), B=c(a=1), C=c(a=1, b=2, c=1), D=c(c=1))
lx.table.bymsets(lsets, as.card=TRUE)
}
\seealso{
\link{lx.tables}
}
