% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.bypairs}
\alias{lx.table.bypairs}
\title{pairwise contingency table(s) from columns}
\usage{
lx.table.bypairs(tab, by = colnames(tab))
}
\arguments{
\item{tab}{a table (matrix or dataframe) of nominal values}

\item{by}{a list of column indices or names (of length k >= 2)}
}
\value{
list of C(k, n) contingency tables crossing all pairs of columns
        in \code{by}
}
\description{
pairwise contingency table(s) from columns
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.table.bypairs(tab)
}
\seealso{
\link{lx.tables}
}
