% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.table.margins}
\alias{lx.table.margins}
\title{add margins (total) to a 2-dimensional contingency table}
\usage{
lx.table.margins(ctab, margin = c(1, 2))
}
\arguments{
\item{ctab}{2-dimensional contingency table}

\item{margin}{which margin to add
1 (total by rows), 
2 (total by columns) 
c(1,2) both margins
0 do nothing}
}
\description{
add margins (total) to a 2-dimensional contingency table
}
\examples{
data(lx.iris)
lx.table.margins(lx.table.bycols(lx.iris, by=1:2))

}
\seealso{
\link{lx.tables}
}
