% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\docType{data}
\name{lx.tables}
\alias{lx.tables}
\title{cross tabulation and contigency tables}
\description{
there are several lx function to built contingency tables.\cr
\itemize{
 \item \link{lx.table}         : count or contingency table with full levels
 \item \link{lx.table.bycols}  : k-dimensional contingency table(s) from columns
 \item \link{lx.table.bypairs} : pairwise contingency table(s) from columns
 \item \link{lx.table.byfacts} : count or contingency table by factors in columns
 \item \link{lx.table.bysets}  : cross table of sets intersections
 \item \link{lx.table.bymsets} : cross table of multisets intersections
 \item \link{lx.table.bycoocs} : cross table of co-occurences in sets
}
\itemize{
 \item \link{lx.table.margins} : add margins (total) to a 2-dimensional contingency table
}
}
