% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.table.r
\name{lx.tapply}
\alias{lx.tapply}
\title{tapply-like with specified levels}
\usage{
lx.tapply(x, by, FUN, levels = NULL, ...)
}
\arguments{
\item{x}{an atomic object, typically a vector}

\item{by}{a vector of factors (or anything that can be converted to factors)
of same length as \code{x}}

\item{FUN}{the function to be applied}

\item{levels}{levels to force}

\item{...}{optional arguments to FUN}
}
\value{
a vector of the same length as \code{levels} (or levels(by) if levels==NULL),
each value is the result of calling FUN on the subset of x at given level.
}
\description{
similar to \link{tapply} but force levels and call FUN
on empty subsets.
}
\examples{
xval  <- c( 1,   2,   3,   0,   1 )   
group <- c('a', 'b', 'c', 'a', 'a')
lx.tapply(xval, by=group, length)
lx.tapply(xval, by=group, sum)
lx.tapply(xval, by=group, length, levels=c('a', 'b', 'd'))

}
\seealso{
\link{lx.tables}
}
