% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.tobase}
\alias{lx.tobase}
\title{get number representation}
\usage{
lx.tobase(num, base = 16L, prefix = "")
}
\arguments{
\item{num}{positive (unsigned) numeric value}

\item{base}{representation base (2 <= base <= 36)}

\item{prefix}{prefix to add (e.g. '0x' for hexadecimal)}
}
\value{
string
}
\description{
get base representation of a positive numeric
(accurate up to 53 bits)
}
\examples{
lx.tobase(123, prefix="0x")  # 0x7b
}
