% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.logging.r
\name{lx.traceback}
\alias{lx.traceback}
\title{print call traceback}
\usage{
lx.traceback(level = "debug", up.frame = 0L)
}
\arguments{
\item{level}{see \link{lx.out}}

\item{up.frame}{see \link{lx.out}}
}
\value{
invisible(traceback list from \link{traceback})
}
\description{
print call traceback on stderr
}
