% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.language.r
\name{lx.true}
\alias{lx.true}
\title{always TRUE function}
\usage{
lx.true(...)
}
\arguments{
\item{...}{anything (ignored arguments)}
}
\value{
TRUE
}
\description{
this function always returns TRUE, whatever
the arguments.
}
