% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.serial.r
\name{lx.unserialize}
\alias{lx.unserialize}
\title{unserialize object from \code{name.rds}}
\usage{
lx.unserialize(name, ...)
}
\arguments{
\item{name}{of object used in \code{lx.serialized}}

\item{...}{additionnal options to \code{\link{readRDS}}}
}
\description{
unserialize object from \code{name.rds}
}
\examples{
data(lx.iris)
tab <- lx.iris
lx.serialize(tab)  # create file tab.Rdata
duptab <- lx.unserialize("tab")
}
\seealso{
\code{\link{lx.serialize}}
}
