% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.algebra.r
\name{lx.upper2vect}
\alias{lx.upper2vect}
\title{get matrix upper triangle as vector}
\usage{
lx.upper2vect(x, diag = FALSE, bycol = TRUE)
}
\arguments{
\item{x}{nxn square numerical matrix}

\item{diag}{logical should the diagonal be included?}

\item{bycol}{logical order result by columns}
}
\value{
numerical vector of matrix upper triangle
}
\description{
get matrix upper triangle as vector
}
\note{
if matrix is not square, a warning is emitted but
result (on the smaller dimension) is still returned
}
\examples{
x <- matrix(1:16, ncol=4)
y <- lx.upper2vect(x, diag=TRUE)
xx <- lx.vect2upper(y, diag=TRUE)
}
\seealso{
\link{upper.tri}, \link{lx.vect2upper}
}
