% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.algebra.r
\name{lx.vect2upper}
\alias{lx.vect2upper}
\title{make matrix from upper triangle vector}
\usage{
lx.vect2upper(x, diag = FALSE, bycol = TRUE, symmetrize = TRUE,
  na.val = NA)
}
\arguments{
\item{x}{numerical vector}

\item{diag}{logical is the diagonal included in x? (see \link{lx.upper2vect})}

\item{bycol}{logical is x ordered by columns? (see \link{lx.upper2vect})}

\item{symmetrize}{logical should result be symmetrized}

\item{na.val}{numerical value for NA (see note)}
}
\value{
numerical matrix
}
\description{
this is the inverse of \link{lx.upper2vect}
}
\note{
undefined values (i.e. diagonal if diag=FALSE and lower matrix if symmetrize=FALSE)
are set to na.val
}
\examples{
x <- matrix(1:16, ncol=4)
y <- lx.upper2vect(x, diag=TRUE)
xx <- lx.vect2upper(y, diag=TRUE)
}
\seealso{
\link{lx.upper2vect}
}
