% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.logging.r
\name{lx.verbose}
\alias{lx.verbose}
\title{set/get verbose level}
\usage{
lx.verbose(level)
}
\arguments{
\item{level}{verbose level (see description)}
}
\value{
in get mode: return current verbose level\cr
in set mode return previous verbose level (i.e. before set)
}
\description{
\bold{set mode:} \code{lx.verbose(level)}
where \code{level} is either an integer value or one of
"debug" (=0); "info" (=10); "warning" (=20); "error" (=100)\cr
\bold{get mode:} \code{lx.verbose()}
}
\examples{
p <- lx.verbose("debug") # set to debug and return previous
lx.verbose(p)            # restore to previous value
}
