% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.logging.r
\name{lx.warn}
\alias{lx.warn}
\title{emit warning}
\usage{
lx.warn(..., level = "warning", up.frame = 0L)
}
\arguments{
\item{...}{anything accepted by \link{lx.out}}

\item{level}{see \link{lx.out}}

\item{up.frame}{see \link{lx.out}}
}
\description{
prints warning on stderr
}
\examples{
lx.warn("pi=", pi, with.mem=TRUE)
}
\seealso{
\link{lx.out}, \link{lx.warnif}, \link{lx.stopif}
}
