% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.logging.r
\name{lx.warnif}
\alias{lx.warnif}
\title{emit warning if condition is true}
\usage{
lx.warnif(condition, ..., level = "warning", up.frame = 0L)
}
\arguments{
\item{condition}{boolean value}

\item{...}{anything accepted by \link{lx.out}}

\item{level}{see \link{lx.out}}

\item{up.frame}{see \link{lx.out}}
}
\value{
invisible(condition)
}
\description{
emit warning if condition is true
}
\examples{
x <- -1
lx.warnif(x < 0, "x=", x, " is negative")
}
\seealso{
\link{lx.out}, \link{lx.warn}, \link{lx.stopif}
}
