% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.wt.mean}
\alias{lx.wt.mean}
\title{weighted mean}
\usage{
lx.wt.mean(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{w}{a numeric vector of (positive) weights}

\item{na.rm}{a logical value indicating whether NA values should
be stripped before the computation proceeds}
}
\value{
weighted mean of x
}
\description{
weighted mean
}
\examples{
lx.wt.mean(1:2, 1:2)

}
