% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.statistics.r
\name{lx.wt.var}
\alias{lx.wt.var}
\title{weighted variance}
\usage{
lx.wt.var(x, w = NULL, as.repeat = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{w}{a numeric vector of (positive) weights}

\item{as.repeat}{a logical value indicating whether weights
should be interpreted as repeated measures or not (see details).}

\item{na.rm}{a logical value indicating whether NA values should
be stripped before the computation proceeds}
}
\value{
weighted variance of x
}
\description{
weighted variance
}
\details{
there is no single accepted formula for the weighted variance. It depends
whether weights should be interpreted as repeats of the same measure (in which
case they usually are integers summing up to the total number of measures) 
(\code{as.repeat=TRUE}) or if thet should be interpreted as a 'confidence'
attached to each measure (\code{as.repeat=FALSE}).\cr
}
\examples{
lx.wt.var(1:2, 1:2)
var(c(1, 2, 2)) # should be the same as previous
lx.wt.var(1:2, 1:2, as.repeat=FALSE) # but this is different
lx.wt.var(1:2, rep(1,2))
var(1:2) # should be the same as previous

}
