% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.fileio.r
\name{print.LXFileHandle}
\alias{print.LXFileHandle}
\title{print method for LXFileHandle}
\usage{
\method{print}{LXFileHandle}(x, ...)
}
\arguments{
\item{x}{file handle (opened by \link{lx.open})}

\item{...}{further arguments passed to or from other methods.}
}
\value{
(invisible) x
}
\description{
print method for LXFileHandle
}
