% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.stack.r
\name{print.Stack}
\alias{print.Stack}
\title{print method for Stack}
\usage{
\method{print}{Stack}(x, ...)
}
\arguments{
\item{x}{Stack}

\item{...}{further arguments passed to or from other methods.}
}
\value{
(invisible) stk
}
\description{
print method for Stack
}
