% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.close}
\alias{tex.close}
\title{close tex report file}
\usage{
tex.close(tex, compile = TRUE, ...)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{compile}{if TRUE, call \link{tex.compile} to produce pdf}

\item{...}{any argument to \link{tex.compile} (see note)}
}
\value{
invisible(tex)
}
\description{
close tex report file
}
\note{
a useful argument is \code{clean=TRUE} that remove
      all auxiliary files created during the conversion
}
