% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.compile}
\alias{tex.compile}
\title{compile tex report file}
\usage{
tex.compile(tex, ...)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{...}{any argument passed to tools::\link{texi2pdf}}
}
\value{
invisible(tex)
}
\description{
compile tex report file
}
