% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.fig.off}
\alias{tex.fig.off}
\title{end tex graphics}
\usage{
tex.fig.off(tex, ...)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.fig.on})}

\item{...}{any attribute to latex:includegraphics}
}
\value{
tex
}
\description{
end tex graphics
}
\note{
it is important to reaffect tex handle upon return
since the internal components have been modified
}
