% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.installed}
\alias{tex.installed}
\title{check if latex is properly installed}
\usage{
tex.installed(cmd = "pdflatex", warn = TRUE)
}
\arguments{
\item{cmd}{latext command to check}

\item{warn}{issue warning on error}
}
\value{
logical
}
\description{
check if latex is properly installed
}
\note{
this function uses \code{system('which command')} and therefore
only works on Unix
}
