% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.open}
\alias{tex.open}
\title{open new tex report file}
\usage{
tex.open(name = "report", title = name, author = lx.sysinfo("user"),
  prolog = NULL)
}
\arguments{
\item{name}{file basename}

\item{title}{report title}

\item{author}{report author}

\item{prolog}{function to add user's specific latex commands
in document prolog. called as \code{prolog(texHandle)}}
}
\value{
tex handle
}
\description{
open new tex report file
}
\note{
the actual tex file is \code{lx.options(tex.dir)/name.tex}
and the final pdf file is \code{name.pdf}
}
