% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.out}
\alias{tex.out}
\title{output line(s) to tex report file}
\usage{
tex.out(tex, ..., lf = FALSE)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{...}{any argument supported by \link{cat}}

\item{lf}{add a latex linefeed \code{\\\\} at end of line}
}
\value{
invisible(tex)
}
\description{
output line(s) to tex report file
}
\seealso{
\link{tex.print} for formatted output
}
