% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.print}
\alias{tex.print}
\title{print formatted argument to tex report file}
\usage{
tex.print(tex, x, ...)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{x}{an object to print (see details)}

\item{...}{any argument supported by \link{print} variant}
}
\value{
invisible(tex)
}
\description{
print formatted argument to tex report file
}
\details{
if \code{x} is of class \code{xtable} then just call the
S3 method \code{print} on \code{x}.\cr
else, first format x using \code{xtable} then call \link{print}

x should be of a type handled by \code{xtable}.
see \link{tex.out} for simple output.
}
\note{
this requires the \code{xtable} package that will be internally
installed if not already present (using internal distribution)
}
\seealso{
\link{tex.out} for simple output
}
