% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.subsection}
\alias{tex.subsection}
\title{start new tex subsection}
\usage{
tex.subsection(tex, name, numbered = TRUE)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{name}{section name}

\item{numbered}{should section be numbered}
}
\value{
invisible(tex)
}
\description{
start new tex subsection
}
