% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx.tex.r
\name{tex.tag}
\alias{tex.tag}
\title{output tex tag to tex report file}
\usage{
tex.tag(tex, tag = NULL, value = NULL, attr = NULL, prepend = FALSE)
}
\arguments{
\item{tex}{handle (as returned by \link{tex.open})}

\item{tag}{tag name}

\item{value}{tag value}

\item{attr}{tag attribute(s)}

\item{prepend}{if TRUE append attributes before value. default is FALSE,
append attributes after value.}
}
\value{
invisible(tex)
}
\description{
output tex tag as\cr
\code{\<tag>{<value>}[<attr>]}
or \code{\<tag>[<attr>]{<value>}} depending upon the \code{prepend} argument.
}
